requirements.txt
Babel==2.14.0
bcrypt==4.1.2
contourpy==1.2.1
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
fonttools==4.51.0
kiwisolver==1.4.5
matplotlib==3.8.4
numpy==1.26.4
packaging==24.0
plotly==5.22.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
six==1.16.0
tenacity==8.2.3
tkcalendar==1.6.1
